requirements.txt
scipy
dash
dash-bio
dash-bootstrap-components
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer
dash-table==5.0.0
numpy==1.21.4
pandas
plotly
scikit-learn
umap-learn
openpyxl
pytest
matplotlib
pyscm-ml@git+https://github.com/aldro61/pyscm.git#egg=pyscm-ml
randomscm@git+https://github.com/thibgo/randomscm.git#egg=randomscm